double -> float for space constraint

+++ pair56_task4/librobot.c
@@ -38,31 +38,31 @@ void turn_pivot_function(int angle) {
     return;
 }
 
-double* distance_wheels_travelled() {
+float* distance_wheels_travelled() {
   int* ticksArr = malloc(2 * sizeof(int));
   drive_getTicks(*ticksArr, *(ticksArr+1));
-  double* distanceWheelsTravelled = malloc(2 * sizeof(double));
+  float* distanceWheelsTravelled = malloc(2 * sizeof(float));
   *distanceWheelsTravelled = (*ticksArr/_ENCODER_CLICKS_PER_FULL_TURN) * 2 * PI * _WHEEL_RADIUS;
   *(distanceWheelsTravelled+1) = (*(ticksArr+1)/_ENCODER_CLICKS_PER_FULL_TURN) * 2 * PI * _WHEEL_RADIUS;
   free(ticksArr);
   return distanceWheelsTravelled;
 }
 
-double angle_change(double* distanceWheelsTravelled) {
+float angle_change(float* distanceWheelsTravelled) {
   return (*distanceWheelsTravelled - *(distanceWheelsTravelled+1)) / _WHEEL_BASE;
 }
 
-double radius_middle(double* distanceWheelsTravelled) {
-  double angleChange = angle_change(distanceWheelsTravelled);
-  double radiusLeft = *distanceWheelsTravelled/angleChange;
-  double radiusRight = *(distanceWheelsTravelled+1)/angleChange;
+float radius_middle(float* distanceWheelsTravelled) {
+  float angleChange = angle_change(distanceWheelsTravelled);
+  float radiusLeft = *distanceWheelsTravelled/angleChange;
+  float radiusRight = *(distanceWheelsTravelled+1)/angleChange;
   return (radiusLeft + radiusRight) / 2;
 }
 
-double* position_change(double* distanceWheelsTravelled, double currentAngle) {
-  double* positionCoordinates = malloc(2 * sizeof(double));
-  double radiusMiddle = radius_middle(distanceWheelsTravelled);
-  double angleChange = angle_change(distanceWheelsTravelled);
+float* position_change(float* distanceWheelsTravelled, float currentAngle) {
+  float* positionCoordinates = malloc(2 * sizeof(float));
+  float radiusMiddle = radius_middle(distanceWheelsTravelled);
+  float angleChange = angle_change(distanceWheelsTravelled);
   *positionCoordinates = radiusMiddle*cos(currentAngle) - radiusMiddle*cos(currentAngle+angleChange);
   *(positionCoordinates + 1) = radiusMiddle*sin(currentAngle+angleChange) - radiusMiddle*sin(currentAngle);
   return positionCoordinates;
